Create workspace packages
