+++ .vscode/settings.json
@@ -6,6 +6,7 @@
         "Deduper",
         "exif",
         "ffprobe",
+        "fileops",
         "Muzaque",
         "muzaques",
         "nihao",
