+++ cli/jest.config.js
