+++ cli/src/fun/umuAscii.ts
