+++ cli/src/operationMap.ts
