+++ cli/src/operations/dateImg.ts
