+++ cli/src/operations/dirTree.ts
