+++ cli/src/operations/global.ts
