+++ cli/src/operations/jpgCompress.ts
