+++ cli/src/operations/nameToTag.ts
