+++ cli/src/operations/operations.helpers.ts
