+++ cli/src/operations/pathAliases.ts
