+++ cli/src/util/colors.ts
