+++ cli/src/util/files.ts
