+++ h-util-ui/Electron/main.ts
@@ -1,12 +1,12 @@
-import { app, BrowserWindow, BrowserWindowConstructorOptions, ipcMain, screen } from "electron";
-import path from "path";
-import { isDev } from "./config";
-import { appConfig } from "./ElectronStore/Configuration";
-import AppUpdater from "./AutoUpdate";
+import { app, BrowserWindow, BrowserWindowConstructorOptions, ipcMain, screen } from 'electron';
+import path from 'path';
+import { isDev } from './config';
+import { appConfig } from './ElectronStore/Configuration';
+import AppUpdater from './AutoUpdate';
 
 async function createWindow() {
     const { width, height } = screen.getPrimaryDisplay().workAreaSize;
-    const appBounds: any = appConfig.get("setting.appBounds");
+    const appBounds: any = appConfig.get('setting.appBounds');
     const BrowserWindowOptions: BrowserWindowConstructorOptions = {
         width: 1200,
         minWidth: 900,
@@ -14,7 +14,7 @@ async function createWindow() {
         minHeight: 600,
 
         webPreferences: {
-            preload: __dirname + "/preload.js",
+            preload: __dirname + '/preload.js',
             devTools: isDev,
         },
         show: false,
@@ -30,9 +30,10 @@ async function createWindow() {
 
     // and load the index.html of the app.
     // win.loadFile("index.html");
-    await mainWindow.loadURL(isDev ? "http://localhost:3000" : `file://${path.join(__dirname, "./index.html")}`);
+    await mainWindow.loadURL(isDev ? 'http://localhost:3000' : `file://${path.join(__dirname, './index.html')}`);
 
-    if (appBounds !== undefined && appBounds !== null && appBounds.width > width && appBounds.height > height) mainWindow.maximize();
+    if (appBounds !== undefined && appBounds !== null && appBounds.width > width && appBounds.height > height)
+        mainWindow.maximize();
     else mainWindow.show();
 
     // this will turn off always on top after opening the application
@@ -44,7 +45,6 @@ async function createWindow() {
     if (isDev) {
         mainWindow.webContents.openDevTools();
     }
-
 
     ipcMain.handle('versions', () => {
         return {
@@ -64,15 +64,15 @@ app.whenReady().then(async () => {
     // if dev
     if (isDev) {
         try {
-            const { installExt } = await import("./installDevTool");
+            const { installExt } = await import('./installDevTool');
             await installExt();
         } catch (e) {
-            console.log("Can not install extension!");
+            console.log('Can not install extension!');
         }
     }
 
     createWindow();
-    app.on("activate", function () {
+    app.on('activate', function () {
         // On macOS it's common to re-create a window in the app when the
         // dock icon is clicked and there are no other windows open.
         if (BrowserWindow.getAllWindows().length === 0) createWindow();
@@ -82,8 +82,8 @@ app.whenReady().then(async () => {
 // Quit when all windows are closed, except on macOS. There, it's common
 // for applications and their menu bar to stay active until the user quits
 // explicitly with Cmd + Q.
-app.on("window-all-closed", () => {
-    if (process.platform !== "darwin") {
+app.on('window-all-closed', () => {
+    if (process.platform !== 'darwin') {
         app.quit();
     }
 });
