+++ h-util-ui/package.json
@@ -46,6 +46,9 @@
     "electron-store": "^8.1.0",
     "electron-updater": "^5.3.0"
   },
+  "workspaces": [
+    "../packages/fileops"
+  ],
   "build": {
     "appId": "com.official-h-util-ui.app",
     "productName": "H-Util UI Nightly",
