+++ package.json
@@ -1,50 +1,26 @@
 {
-  "name": "hoarder-util",
+  "name": "h-util-monorepo",
   "version": "1.0.0",
-  "description": "",
+  "description": "Turns some common data operations into a CLI tool, to prevent needing various third party apps and utilities and configs. This includes:",
   "main": "index.js",
   "scripts": {
-    "test-install": "npm run update && h-util nihao",
-    "test": "jest --verbose",
-    "test:watch": "jest --verbose --watch",
-    "start": "nodemon --watch 'src/**/*.ts' --exec 'ts-node' index.ts",
-    "ts-n": "ts-node index.ts",
-    "build": "tsc -p .",
-    "local": "npm i -g",
-    "refresh": "rm -rf ./node_modules ./package-lock.json && npm install",
-    "update": "npm run build && npm run local",
-    "ex": "npm run ts-n -- op-alias ls"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "tsc --build --verbose",
+    "build:common": "npm run build --workspace ./packages/common",
+    "build:fileops": "npm run build --workspace ./packages/fileops"
   },
-  "bin": {
-    "h-util": "./lib/index.js"
-  },
-  "keywords": [],
   "author": "",
   "license": "ISC",
+  "workspaces": [
+    "packages/*",
+    "packages/common",
+    "packages/fileops"
+  ],
   "dependencies": {
-    "cli-progress": "^3.12.0",
-    "colors": "^1.4.0",
-    "commander": "^11.1.0",
-    "conf": "^10.2.0",
-    "dateformat": "^4.6.3",
-    "exif": "^0.6.0",
-    "figlet": "^1.7.0",
-    "fluent-ffmpeg": "^2.1.2",
-    "is-invalid-path": "^1.0.2",
-    "jimp": "^0.22.10",
-    "prompt-sync": "^4.2.0",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.3.3"
+    "typescript": "^5.5.4"
   },
   "devDependencies": {
-    "@types/cli-progress": "^3.11.5",
-    "@types/dateformat": "^3.0.1",
-    "@types/figlet": "^1.5.8",
-    "@types/fluent-ffmpeg": "^2.1.24",
-    "@types/jest": "^29.5.11",
-    "@types/node": "^20.10.4",
-    "@types/prompt-sync": "^4.2.3",
-    "jest": "^29.7.0",
-    "ts-jest": "^29.1.1"
+    "@types/node": "^20.14.12",
+    "ts-node": "^10.9.2"
   }
 }
