+++ packages/fileops/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@common/fileops",
+  "version": "1.0.0",
+  "main": "build/index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "tsc --build"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "description": "",
+  "dependencies": {
+    "@common/common": "^1.0.0"
+  }
+}
