+++ packages/fileops/src/index.ts
@@ -1,0 +1,3 @@
+import { printf } from '@common/common';
+
+export const printfExtended = (message: string) => printf(message);
