+++ tsconfig.json
@@ -1,13 +1,25 @@
 {
-     "compilerOptions": {
-          "typeRoots": ["node_modules/@types"],
-          "outDir": "lib",
-          "module": "CommonJS",
-          "target": "es2017"
-     },
-     "rootDir": "./",
-     "baseUrl": "./",
-     "paths": {
-          "*": ["./*"]
-     },
+  "compilerOptions": {
+    "incremental": true,
+    "target": "es2022",
+    "module": "commonjs",
+    "declaration": true,
+    "strict": true,
+    "moduleResolution": "node",
+    "esModuleInterop": true,
+    "skipLibCheck": true,
+    "forceConsistentCasingInFileNames": true,
+    "rootDir": "./src",
+    "outDir": "./build"
+  },
+  "files": [
+  ],
+  "references": [
+    {
+      "path": "./packages/common"
+    },
+    {
+      "path": "./packages/fileops"
+    }
+  ]
 }
